youtube_video_summarizer/requirements.txt
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_huggingface==0.3.0
langchain_ollama==0.3.3
langchain_openai==0.3.27
python-dotenv==1.1.1
whisper==1.1.10
yt_dlp==2025.6.30
